SwingConsolePane: keep console pane at the bottom
If the console pane is scrolled to the bottom, and there is new output,
let's keep it scrolled to the bottom, so that output can be easily
tracked by the human eye. But if the scroll bar is _not_ at the bottom,
let's leave it where it is, to avoid disrupting the user.

This matches the behavior of most terminal applications.

Closes #7.