Fix non-primitive numeric type default values
This change leans on the new ModuleService#getDefaultValue(ModuleItem)
method, which is smarter about default value computation across the
board. It avoids the pitfall of bounding min/max using the widget
model's min and max -- which fall back to the numeric type's min/max
values for the purposes of bounding the widget -- and instead uses the
ModuleItem's stated min/max, or a default converted from "0" if none.