Add focus listener to JSpinner text
This (mostly) ensures the text of a JSpinner is selected on focus. On OSX it
seems to always work with tabs, and mostly work with mouse clicking. However
there does seem to be a race condition where the text selection can be lost,
presumably from the JSPinner firing a validation event after the text field's
selection has been processed.

See https://github.com/scijava/scijava-ui-swing/issues/12 for future work.