Revert "SwingConsolePane: first cut at batching the output"
This reverts commit 881065a3f649e197a8cf5536db85e607ed9c70ea.

Batching the output does not provide great performance gains, and it
causes problems with the reliability of output delivery. But we'll
leave the work in place in the history: a) so the history documents
why things are the way they are; and b) in case anyone wants to make
a more nuanced attempt at batching output in the future.