OptionsLookAndFeel: Support for FlatLaf
This implements support for FlatLaf Look and Feels.
A static method is also included so that Look and Feel can be set at
startup. E.g., for Fjii the following call in the StartupMacros could be
used to set the Look and Feel programmatically at startup:

    eval("js","importClass(Packages.org.scijava.ui.swing.options.OptionsLookAndFeel);OptionsLookAndFeel.setupLookAndFeel('FlatLaf Dark');")

NB: there are currently several hardwired foreground/background colors in several classes (e.g., ConsolePanel) that will interfere with dark themes.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>