Adds swing ui which monitors scijava's tasks
Adds a circular progress bar component to the status bar which  displays
the overall progress of all registered tasks created via the scijava
task service. When the user clicks on the circular progress bar, a frame
containing a table summarizing each task progression is displayed.

Each task is rendered in the table with a custom renderer, displaying a
label with the name of the task and its status, a linear progress bar,
and, if enabled, a label with an estimation of the remaining time until
task completion. The estimation assumes a constant completion speed
since the start of the task registration. Additionally, the second
column of the table displays a clickable stop icon and which can be used
to cancel the task.

Before cancellation, a user confirmation dialog can be enabled or
disabled.