SwingConsolePane: detect infinite loops
When initializing, we invoke `new ConsolePanel`. If that invocation
triggers more output to the console, we will get stuck in a loop.
Let's be proactive about that situation and fail fast instead.

This can happen e.g. if the chosen LookAndFeel is broken due to
missing or inaccessible ComponentUI classes.