SwingObjectWidget: set values by string comparison
Previously, we were trying to set the JComboBox value
with a new instance of the object.

We don't always have the possibility to keep a list of the allowed
choices for the current model, nor can we override the `equals()` method
for the `Object`s of this widget instance.
So we have to fall back to comparing all combobox items to the current
value by their `toString()` and `ObjectService.getName()` strings.